backend/ksap-backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
supabase==2.3.4
python-dotenv==1.1.1
requests==2.31.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0
redis==6.2.0
celery==5.5.3
psycopg2-binary==2.9.10

